Alterações de relacionamentos

+++ Controllers/ProdutosController.cs
@@ -52,8 +52,23 @@ namespace OdontoBoxApi.Controllers;
             {
                 return Problem("Entidade Produtos é nula.");
             }
+            
+            //add coisa aqui kaua, validação se o fornecedor existe
+            var fornecedorExiste = await _context.Fornecedores.FindAsync(produto.FornecedorId);
+            if (fornecedorExiste == null)
+            {
+                return BadRequest($"Fornecedor com ID: {produto.FornecedorId} não encontrado.");
+            }
+            //validar se o nome n ta vazio
+            if (string.IsNullOrWhiteSpace(produto.Nome))
+            {
+                    return BadRequest("O nome do produto é obrigatório.");
+            }
+                    
+            
             _context.Produtos.Add(produto);
             await _context.SaveChangesAsync();
+            
             
             return CreatedAtAction("GetProdutoId", new { id = produto.Id }, produto); 
         }
@@ -88,23 +103,33 @@ namespace OdontoBoxApi.Controllers;
         public async Task<ActionResult> DeleteProduto(int id)
         {
            if (_context.Produtos == null)
-        {
-            return Problem();
-        }
-        var produto = await _context.Produtos.FindAsync(id);
-        if (produto == null) 
-        {
-            return NotFound();
-        }
+            {
+                return Problem();
+            }
+            var produto = await _context.Produtos.FindAsync(id);
+            if (produto == null) 
+            {
+                return NotFound();
+            }
+            
+            //validacao de saida pra poder deletar, se tiver saida pra esse produto, n deixa deletar
+            var saidasAssociadas = await _context.Saidas
+                .Where(s => s.ProdutoId == id)
+                .CountAsync();
 
-        _context.Produtos.Remove(produto);
-        await _context.SaveChangesAsync();
+            if (saidasAssociadas > 0)
+            {
+                return BadRequest($"Não é possível deletar o produto pois existem {saidasAssociadas} saída(s) associada(s) a ele.");
+            }
 
-        return NoContent();
+            _context.Produtos.Remove(produto);
+            await _context.SaveChangesAsync();
+
+            return NoContent();
         }
-         private bool ProdutosExists(int id)
-    {
-        return (_context.Produtos?.Any(e => e.Id == id)).GetValueOrDefault();
-    }
+        private bool ProdutosExists(int id)
+        {
+            return (_context.Produtos?.Any(e => e.Id == id)).GetValueOrDefault();
+        }
     }
